refactor convergence criteria
