+++ problems/n_dim/vdp_beam.py
@@ -1,5 +1,6 @@
 import matplotlib.pyplot as plt
 import numpy as np
+
 from sao.problems.problem import Problem
