+++ problems/one_dim/polynomial_1d.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from sao.problems import Problem
