+++ problems/svanberg1987.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from sao.problems.problem import Problem
 
 
@@ -137,6 +138,7 @@ class TwoBarTruss(Problem):
 
 if __name__ == "__main__":
     from problems.util.fd import finite_difference
+
     dx = 1e-7
     problem = TwoBarTruss()
     finite_difference(problem, problem.x0 + 1, dx)
