+++ problems/topology_optimization/stress.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 import util.to_utils as utils
 from sao.problems import Problem
