+++ sao/convergence_criteria/criteria.py
@@ -1,9 +1,7 @@
 """Abstract implementation for convergence criteria."""
 import abc
-import math
 import operator
-
-import numpy as np
+from change import ObjectiveChange
 
 
 class Criterion(abc.ABC):
@@ -98,141 +96,3 @@ class Criteria(Criterion):
         return self.operator(bool(self.left), bool(self.right))
 
 
-class ObjectiveChange(Criterion):
-    """Keeps track of the relative objective changes between iterations."""
-
-    def __init__(self, objective, tolerance=1e-4, normalise=False):
-        super().__init__()
-        self.objective = objective
-        self.tolerance = tolerance
-        self.previous = np.full_like(self.objective, np.inf)
-        self.normalise = normalise
-        self.value = None
-
-    def __call__(self):
-        """Evaluate the objective changes between iterations."""
-        current = self.objective
-
-        self.value = abs(current - self.previous)
-        if self.normalise:
-            # If previous is `np.inf`, the change is already set to infinite,
-            # and the normalisation can be skipped to avoid division by `Inf`.
-            if not np.isinf(self.previous):
-                self.value /= abs(self.previous)
-
-        self.done = bool(self.value < self.tolerance)
-
-        # make sure to keep track of the previous iterations
-        self.previous = current.copy()
-
-
-class VariableChange(Criterion):
-    """Enforces an absolute, maximum variable change on all design variables.
-
-    The variable change is determined between two consecutive iterations and
-    the tolerance should be achieved for all defined design variables. If
-    desired, the variable change can be scaled, for instance to normalise the
-    change with respect to the maximum range of the design variables.
-    """
-
-    def __init__(self, variables, tolerance=1e-4, scaling=1.0):
-        """Initialise the criteria with a tolerance and scaling"""
-        super().__init__()
-        self.variables = variables
-        self.tolerance = tolerance
-        self.scaling = scaling
-        self.previous = math.inf
-
-    def __call__(self):
-        """Assert all variables show a sufficiently small change."""
-        current = self.variables
-
-        # The (scaled) change of all variables should be sufficiently small
-        # before the variable change is satisfied.
-        self.done = all(
-            abs((current - self.previous) / self.scaling) < self.tolerance)
-
-        # keep track of the previous iterations value of the variables
-        self.previous = current.copy()
-
-
-class VariableChangeNorm(Criterion):
-    """Enforces an absolute, maximum variable change on all design variables.
-
-    The variable change is determined between two consecutive iterations and
-    the tolerance should be achieved for all defined design variables. If
-    desired, the variable change can be scaled, for instance to normalise the
-    change with respect to the maximum range of the design variables.
-    """
-
-    def __init__(self, variables, tolerance=1e-4, scaling=1.0):
-        """Initialise the criteria with a tolerance and scaling"""
-        super().__init__()
-        self.variables = variables
-        self.tolerance = tolerance
-        self.scaling = scaling
-        self.previous = math.inf
-        self.value = None
-
-    def __call__(self):
-        """Assert the norm of the variables' change is sufficiently small."""
-        current = self.variables
-
-        # The (scaled) change of all variables should be sufficiently small
-        # before the variable change is satisfied.
-        self.value = np.linalg.norm((current - self.previous) / self.scaling)
-        self.done = (self.value < self.tolerance).astype(bool)
-
-        # keep track of the previous iterations value of the variables
-        self.previous = current.copy()
-
-
-class Feasibility(Criterion):
-    """Enforces feasibility of all constraints with some "slack"."""
-
-    def __init__(self, contraints, slack=1e-4):
-        """Initialise the feasibility criteria with some allowed "slack".
-
-        The slack variable can be provided as a single value or as a list of
-        value, one per available constraint function. When only a single value
-        is provided, that value is considered for all constraint functions.
-        """
-        super().__init__()
-        self.constraints = contraints
-
-        # Test if the provided slack variables allow iteration, if not, the
-        # single value is repeated for the required number of constraints.
-        try:
-            _ = iter(slack)
-            self.slack = slack
-        except TypeError:
-            self.slack = [slack] * len(self.constraints)
-
-        err_msg = f"Wrong number of slack variables: {len(self.slack)}"
-        assert len(self.slack) == len(self.constraints), err_msg
-
-    def __call__(self):
-        """Assert feasibility for each constraint with its allowed slack."""
-        self.done = all(c < s
-                        for (c, s) in zip(self.constraints, self.slack))
-
-
-class IterationCount(Criterion):
-    """Enforces a maximum number of iterations.
-
-    This keeps track of an internal counter that increments on each evaluation
-    of ``__call__``. Thus, when setting up a loop as follows, this criterion
-    will ensure at most 50 iterations are performed.
-
-    >>> convergence = IterationCount(50):
-    >>> while not convergence
-    """
-
-    def __init__(self, max_iteration=50):
-        super().__init__()
-        self.iteration = 0
-        self.max_iteration = max_iteration
-
-    def __call__(self):
-        self.iteration += 1
-        self.done = self.iteration >= self.max_iteration
