+++ sao/convergence_criteria/kkt.py
@@ -1,15 +1,14 @@
 import numpy as np
+from .criteria import Criterion
 
-from .convergence_criterion import ConvergenceCriterion
 
-
-class KKT(ConvergenceCriterion):
+class KKT(Criterion):
     """
     The KKT conditions as a convergence criterion.
     """
 
     def __init__(self, **kwargs):
-        super().__init__(**kwargs)
+        super().__init__()
         self.xmin = kwargs.get('x_min', None)
         self.xmax = kwargs.get('x_max', None)
         self.value = None  # For logging/plotting purposes
