+++ sao/util/__init__.py
@@ -1,2 +1,1 @@
-from .plotter import Plot
 from .tools import parse_to_list, fill_set_when_emtpy
